+++ src/classes/Button.ts
@@ -1,4 +1,4 @@
-import { ButtonInteraction, CacheType } from 'discord.js';
+import { ButtonInteraction } from 'discord.js';
 
 import IButton from '../interfaces/IButton';
 import Client from './Client';
@@ -13,6 +13,7 @@ export default class Button implements IButton {
     this.id = id;
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  Execute(interaction: ButtonInteraction<CacheType>): void {}
+  Execute(interaction: ButtonInteraction): void {
+    throw new Error(`Method not implemented in ${interaction.id} button.`);
+  }
 }