+++ src/commands/image/kiss.ts
@@ -1,5 +1,5 @@
 import Client from '../../classes/Client';
-import FunnyImageCommand from '../../classes/FunnyImage';
+import FunnyImageCommand from '../../classes/commands/FunnyImage';
 import FunnyImageEffect from '../../libs/funny-image-effect';
 
 export default class Kiss extends FunnyImageCommand {