+++ src/commands/music/filter.ts
@@ -1,21 +1,23 @@
 import {
   ApplicationCommandOptionType,
   ChatInputCommandInteraction,
+  Guild,
   GuildMember,
+  GuildTextBasedChannel,
   PermissionsBitField,
 } from 'discord.js';
+import { KazagumoPlayer } from 'kazagumo';
 import { capitalize, find, map } from 'lodash';
 
 import Client from '../../classes/Client';
-import Command from '../../classes/Command';
+import MusicCommand from '../../classes/commands/Music';
 import Category from '../../enums/Category';
-import DefaultEmbed, { ErrorEmbed, WarningEmbed } from '../../libs/discord-embeds';
+import DefaultEmbed, { WarningEmbed } from '../../libs/discord-embeds';
 import EmbedTitles from '../../libs/embed-titles';
-import env from '../../libs/env';
 import audioEffects from '../../libs/filters';
 import MusicControllerUpdate from '../../libs/music-controller-update';
 
-export default class Filter extends Command {
+export default class Filter extends MusicCommand {
   constructor(client: Client) {
     super(client, {
       name: 'filter',
@@ -38,74 +40,31 @@ export default class Filter extends Command {
     });
   }
 
-  async Execute(interaction: ChatInputCommandInteraction) {
-    const { guild, options, channel } = interaction;
-    const member = interaction.member as GuildMember | null;
+  async MusicCommandExecute(interaction: ChatInputCommandInteraction) {
+    const guild = interaction.guild as Guild;
+    const member = interaction.member as GuildMember;
+    const channel = interaction.channel as GuildTextBasedChannel;
+    const bot = guild.members.me as GuildMember;
 
-    const filter = options.getString('filter', true);
+    const filter = interaction.options.getString('filter', true);
 
-    if (!guild || !member || !channel) {
-      interaction.reply({
-        embeds: [ErrorEmbed(this.client, EmbedTitles.music, 'Помилка обробки команди')],
-        ephemeral: true,
-      });
-      return;
-    }
+    const musicChannelId = this.client.GetSetting('music_channel_id');
 
-    if (channel?.id !== env.MUSIC_CHANNEL_ID) {
-      interaction.reply({
-        embeds: [
-          ErrorEmbed(
-            this.client,
-            EmbedTitles.music,
-            `Ви можете використовувати цю команду тільки в ${this.client.channels.cache.get(env.MUSIC_CHANNEL_ID)}`
-          ),
-        ],
-        ephemeral: true,
-      });
-      return;
-    }
-
-    if (!member?.voice.channel) {
-      interaction.reply({
-        embeds: [ErrorEmbed(this.client, EmbedTitles.music, 'Ви не в голосовому каналі')],
-        ephemeral: true,
-      });
-      return;
-    }
-
-    if (guild?.members.me?.voice.channelId && member?.voice.channelId !== guild?.members.me?.voice.channelId) {
-      interaction.reply({
-        embeds: [
-          ErrorEmbed(
-            this.client,
-            EmbedTitles.music,
-            `Ви повинні бути в голосовому каналі разом з ботом (${guild.members.me.voice})`
-          ),
-        ],
-        ephemeral: true,
-      });
-      return;
-    }
+    await this.NullCheck(interaction);
+    await this.MusicChannelCheck(channel.id, musicChannelId);
+    await this.UserVoiceChannelCheck(member, bot);
 
     await interaction.deferReply({ ephemeral: true });
-    const player = this.client.lavalink.players.get(guild!.id);
+    const player = this.client.lavalink.players.get(guild!.id) as KazagumoPlayer;
 
-    if (!player || !player.queue || !player.queue.current) {
-      interaction.editReply({ embeds: [WarningEmbed(this.client, EmbedTitles.music, 'Наразі черга пуста.')] });
-      return;
-    }
-
+    await this.ClearQueueCheck(player);
     const filterObject = find(audioEffects, ['key', filter]);
-
     if (!filterObject) {
       interaction.editReply({ embeds: [WarningEmbed(this.client, EmbedTitles.music, 'Не вдалось знайти фільтр')] });
       return;
     }
-
     await player.shoukaku.setFilters({ ...filterObject.value, volume: player.volume });
-    await MusicControllerUpdate(this.client, player, player.queue.current);
-
+    await MusicControllerUpdate(this.client, player, player.queue.current!);
     const embed = DefaultEmbed(this.client)
       .setTitle(EmbedTitles.music)
       .setDescription(`🎶 Змінено фільтр на ${filter}`);