+++ src/commands/setup/setup.ts
@@ -12,21 +12,73 @@ export default class Setup extends Command {
       category: Category.Setup,
       options: [
         {
-          name: 'temp-voice',
-          description: 'Setup temp voice channels module',
+          name: 'show',
+          description: 'Show all setup options',
+          type: ApplicationCommandOptionType.Subcommand,
+        },
+        {
+          name: 'music-channel',
+          description: 'Setup music channel',
           type: ApplicationCommandOptionType.Subcommand,
           options: [
             {
-              name: 'join-to-talk-channel',
-              description: 'Join to talk channel. Must be a voice channel',
+              name: 'channel',
+              description: 'Text channel for msuic module',
               type: ApplicationCommandOptionType.Channel,
-              required: true,
+              required: false,
             },
+          ],
+        },
+        {
+          name: 'global-channel',
+          description: 'Setup global channel',
+          type: ApplicationCommandOptionType.Subcommand,
+          options: [
             {
-              name: 'temp-voice-channels-category',
-              description: 'Temp voice channels category. Must be a category',
+              name: 'channel',
+              description: 'Base text channel',
               type: ApplicationCommandOptionType.Channel,
-              required: true,
+              required: false,
+            },
+          ],
+        },
+        {
+          name: 'join-role',
+          description: 'Setup join role',
+          type: ApplicationCommandOptionType.Subcommand,
+          options: [
+            {
+              name: 'role',
+              description: 'Role give to new members',
+              type: ApplicationCommandOptionType.Role,
+              required: false,
+            },
+          ],
+        },
+        {
+          name: 'join-to-talk-channel',
+          description: 'Setup join to talk channel',
+          type: ApplicationCommandOptionType.Subcommand,
+          options: [
+            {
+              name: 'channel',
+              description: 'Voice channel for join to talk module',
+              type: ApplicationCommandOptionType.Channel,
+              required: false,
+            },
+          ],
+        },
+
+        {
+          name: 'temp_voice_channels_category',
+          description: 'Setup temp voice channels category',
+          type: ApplicationCommandOptionType.Subcommand,
+          options: [
+            {
+              name: 'category',
+              description: 'Category for temp voice channels',
+              type: ApplicationCommandOptionType.Channel,
+              required: false,
             },
           ],
         },